smolderon & tindral updates
standard opener, prevent the intermission short window;
show coldblood if shadow blades isn't up. this is for intermissions to align it with ashes where CB sec tech will be used before blades is pulled with vanish;
allow flag to be used without symbol syncing on tindral to allow an extra use on pull;
sync shuriken tornado to dance;
sync dance to flag this is so dance is suggested after the extra flag is used.